requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
pydantic==2.12.5
pydantic[email]==2.12.5
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
alembic==1.13.1
